+++ src/components/editor/PageEditor/usePageEditorEffects.ts
@@ -60,7 +60,7 @@ export function usePageEditorEffects(options: UsePageEditorEffectsOptions) {
     toast,
   } = options;
 
-  const { setPageContext } = useAIChatContext();
+  const { setPageContext, contentAppendHandlerRef } = useAIChatContext();
 
   // /page/new への直接アクセスはホームへリダイレクト
   useEffect(() => {
@@ -156,4 +156,15 @@ export function usePageEditorEffects(options: UsePageEditorEffectsOptions) {
     }
     return () => setPageContext(null);
   }, [title, currentPageId, content, setPageContext]);
+
+  // AI追記時にエディタ内容を同期するハンドラを登録
+  useEffect(() => {
+    if (currentPageId) {
+      contentAppendHandlerRef.current = setContent;
+      return () => {
+        contentAppendHandlerRef.current = null;
+      };
+    }
+    return undefined;
+  }, [currentPageId, setContent, contentAppendHandlerRef]);
 }
