+++ src/lib/aiChatActionHelpers.ts
@@ -1,35 +1,13 @@
 import type { PageSummary } from "@/types/page";
 import { MAX_REFERENCED_PAGES, type ReferencedPage } from "@/types/aiChat";
 import { extractWikiLinksFromContent } from "./wikiLinkUtils";
+import { convertMarkdownToTiptapContent } from "./markdownToTiptap";
 
-type TiptapMark = {
-  type: string;
-  attrs?: Record<string, unknown>;
-};
-
-type TiptapTextNode = {
-  type: "text";
-  text?: string;
-  marks?: TiptapMark[];
-};
-
-type TiptapBlockNode = {
-  type: string;
-  attrs?: Record<string, unknown>;
-  content?: Array<TiptapBlockNode | TiptapTextNode>;
-};
+export { convertMarkdownToTiptapContent } from "./markdownToTiptap";
 
 type TiptapDoc = {
   type: "doc";
-  content: TiptapBlockNode[];
-};
-
-type MatchInfo = {
-  index: number;
-  length: number;
-  text: string;
-  type: "wikiLink" | "link" | "bold" | "italic";
-  url?: string;
+  content: Array<{ type: string; attrs?: Record<string, unknown>; content?: unknown[] }>;
 };
 
 function escapeRegExp(value: string): string {
@@ -55,201 +33,6 @@ function parseTiptapDoc(content: string): TiptapDoc | null {
   } catch {
     return null;
   }
-}
-
-function parseInlineContent(text: string): TiptapTextNode[] {
-  const content: TiptapTextNode[] = [];
-  const matches: MatchInfo[] = [];
-
-  const wikiLinkRegex = /\[\[([^\]]+)\]\]/g;
-  let match: RegExpExecArray | null;
-  while ((match = wikiLinkRegex.exec(text)) !== null) {
-    matches.push({
-      index: match.index,
-      length: match[0].length,
-      text: match[1],
-      type: "wikiLink",
-    });
-  }
-
-  const externalLinkRegex = /(?<!\[)\[([^[\]]+)\]\(([^)]+)\)/g;
-  while ((match = externalLinkRegex.exec(text)) !== null) {
-    matches.push({
-      index: match.index,
-      length: match[0].length,
-      text: match[1],
-      url: match[2],
-      type: "link",
-    });
-  }
-
-  const boldRegex = /\*\*([^*]+)\*\*/g;
-  while ((match = boldRegex.exec(text)) !== null) {
-    matches.push({
-      index: match.index,
-      length: match[0].length,
-      text: match[1],
-      type: "bold",
-    });
-  }
-
-  const italicRegex = /(?<!\*)\*([^*]+)\*(?!\*)/g;
-  while ((match = italicRegex.exec(text)) !== null) {
-    matches.push({
-      index: match.index,
-      length: match[0].length,
-      text: match[1],
-      type: "italic",
-    });
-  }
-
-  matches.sort((a, b) => a.index - b.index);
-
-  let lastIndex = 0;
-  for (const item of matches) {
-    if (item.index < lastIndex) continue;
-
-    if (item.index > lastIndex) {
-      content.push({
-        type: "text",
-        text: text.slice(lastIndex, item.index),
-      });
-    }
-
-    if (item.type === "wikiLink") {
-      content.push({
-        type: "text",
-        text: `[[${item.text}]]`,
-        marks: [
-          {
-            type: "wikiLink",
-            attrs: {
-              title: item.text,
-              exists: false,
-            },
-          },
-        ],
-      });
-    } else if (item.type === "link") {
-      content.push({
-        type: "text",
-        text: item.text,
-        marks: [
-          {
-            type: "link",
-            attrs: {
-              href: item.url,
-              target: "_blank",
-              rel: "noopener noreferrer",
-            },
-          },
-        ],
-      });
-    } else if (item.type === "bold") {
-      content.push({
-        type: "text",
-        text: item.text,
-        marks: [{ type: "bold" }],
-      });
-    } else if (item.type === "italic") {
-      content.push({
-        type: "text",
-        text: item.text,
-        marks: [{ type: "italic" }],
-      });
-    }
-
-    lastIndex = item.index + item.length;
-  }
-
-  if (lastIndex < text.length) {
-    content.push({
-      type: "text",
-      text: text.slice(lastIndex),
-    });
-  }
-
-  if (content.length === 0 && text.length > 0) {
-    content.push({
-      type: "text",
-      text,
-    });
-  }
-
-  return content;
-}
-
-export function convertMarkdownToTiptapContent(markdown: string): string {
-  const lines = markdown.split("\n");
-  const doc: TiptapDoc = {
-    type: "doc",
-    content: [],
-  };
-
-  for (const line of lines) {
-    if (line.trim() === "") {
-      doc.content.push({ type: "paragraph" });
-      continue;
-    }
-
-    if (line.startsWith("### ")) {
-      doc.content.push({
-        type: "heading",
-        attrs: { level: 3 },
-        content: parseInlineContent(line.slice(4)),
-      });
-      continue;
-    }
-
-    if (line.startsWith("## ")) {
-      doc.content.push({
-        type: "heading",
-        attrs: { level: 2 },
-        content: parseInlineContent(line.slice(3)),
-      });
-      continue;
-    }
-
-    if (line.startsWith("# ")) {
-      doc.content.push({
-        type: "heading",
-        attrs: { level: 1 },
-        content: parseInlineContent(line.slice(2)),
-      });
-      continue;
-    }
-
-    if (line.startsWith("- ") || line.startsWith("* ")) {
-      const listItem: TiptapBlockNode = {
-        type: "listItem",
-        content: [
-          {
-            type: "paragraph",
-            content: parseInlineContent(line.slice(2)),
-          },
-        ],
-      };
-
-      const lastNode = doc.content[doc.content.length - 1];
-      if (lastNode?.type === "bulletList") {
-        if (!lastNode.content) lastNode.content = [];
-        lastNode.content.push(listItem);
-      } else {
-        doc.content.push({
-          type: "bulletList",
-          content: [listItem],
-        });
-      }
-      continue;
-    }
-
-    doc.content.push({
-      type: "paragraph",
-      content: parseInlineContent(line),
-    });
-  }
-
-  return JSON.stringify(doc);
 }
 
 export function appendTiptapContent(existingContent: string, appendedContent: string): string {
