+++ src/lib/wikiGenerator.ts
@@ -417,241 +417,4 @@ export async function generateWikiContentStream(
   }
 }
 
-/**
- * MarkdownをTiptap JSON形式に変換するためのヘルパー
- * WikiLinkを適切なマークに変換
- */
-export function convertMarkdownToTiptapContent(markdown: string): string {
-  // TiptapのsetContent()はHTMLまたはJSONを受け付ける
-  // ここではシンプルにMarkdownテキストをTiptapの段落構造に変換
-
-  const lines = markdown.split("\n");
-  const doc: {
-    type: string;
-    content: Array<{
-      type: string;
-      attrs?: Record<string, unknown>;
-      content?: Array<{
-        type: string;
-        text?: string;
-        marks?: Array<{ type: string; attrs?: Record<string, unknown> }>;
-      }>;
-    }>;
-  } = {
-    type: "doc",
-    content: [],
-  };
-
-  for (const line of lines) {
-    if (line.trim() === "") {
-      // 空行
-      doc.content.push({ type: "paragraph" });
-    } else if (line.startsWith("### ")) {
-      // H3
-      doc.content.push({
-        type: "heading",
-        attrs: { level: 3 },
-        content: parseInlineContent(line.slice(4)),
-      });
-    } else if (line.startsWith("## ")) {
-      // H2
-      doc.content.push({
-        type: "heading",
-        attrs: { level: 2 },
-        content: parseInlineContent(line.slice(3)),
-      });
-    } else if (line.startsWith("# ")) {
-      // H1
-      doc.content.push({
-        type: "heading",
-        attrs: { level: 1 },
-        content: parseInlineContent(line.slice(2)),
-      });
-    } else if (line.startsWith("- ") || line.startsWith("* ")) {
-      // リストアイテム（簡易実装）
-      const existingList = doc.content[doc.content.length - 1];
-      const listItem = {
-        type: "listItem",
-        content: [
-          {
-            type: "paragraph",
-            content: parseInlineContent(line.slice(2)),
-          },
-        ],
-      };
-
-      if (existingList && existingList.type === "bulletList") {
-        (existingList.content as Array<typeof listItem>).push(listItem);
-      } else {
-        doc.content.push({
-          type: "bulletList",
-          content: [listItem] as typeof doc.content,
-        });
-      }
-    } else {
-      // 通常の段落
-      doc.content.push({
-        type: "paragraph",
-        content: parseInlineContent(line),
-      });
-    }
-  }
-
-  return JSON.stringify(doc);
-}
-
-/**
- * インラインコンテンツをパース（WikiLink、外部リンク、太字、斜体など）
- */
-function parseInlineContent(text: string): Array<{
-  type: string;
-  text?: string;
-  marks?: Array<{ type: string; attrs?: Record<string, unknown> }>;
-}> {
-  const content: Array<{
-    type: string;
-    text?: string;
-    marks?: Array<{ type: string; attrs?: Record<string, unknown> }>;
-  }> = [];
-
-  // 全てのマッチを収集
-  interface MatchInfo {
-    index: number;
-    length: number;
-    text: string;
-    type: string;
-    url?: string;
-  }
-
-  const matches: MatchInfo[] = [];
-
-  let match;
-
-  // WikiLink
-  const wikiLinkRegex = /\[\[([^\]]+)\]\]/g;
-  while ((match = wikiLinkRegex.exec(text)) !== null) {
-    matches.push({
-      index: match.index,
-      length: match[0].length,
-      text: match[1],
-      type: "wikiLink",
-    });
-  }
-
-  // 外部リンク [テキスト](URL) - WikiLinkと区別するため、[[を含まないものだけ
-  const externalLinkRegex = /(?<!\[)\[([^[\]]+)\]\(([^)]+)\)/g;
-  while ((match = externalLinkRegex.exec(text)) !== null) {
-    matches.push({
-      index: match.index,
-      length: match[0].length,
-      text: match[1],
-      url: match[2],
-      type: "link",
-    });
-  }
-
-  // 太字
-  const boldRegex = /\*\*([^*]+)\*\*/g;
-  while ((match = boldRegex.exec(text)) !== null) {
-    matches.push({
-      index: match.index,
-      length: match[0].length,
-      text: match[1],
-      type: "bold",
-    });
-  }
-
-  // 斜体（太字とは別に）
-  const italicRegex = /(?<!\*)\*([^*]+)\*(?!\*)/g;
-  while ((match = italicRegex.exec(text)) !== null) {
-    matches.push({
-      index: match.index,
-      length: match[0].length,
-      text: match[1],
-      type: "italic",
-    });
-  }
-
-  // インデックスでソート
-  matches.sort((a, b) => a.index - b.index);
-
-  // テキストを分割して処理
-  let lastIndex = 0;
-  for (const m of matches) {
-    // 重複するマッチをスキップ
-    if (m.index < lastIndex) continue;
-
-    // マッチ前のプレーンテキスト
-    if (m.index > lastIndex) {
-      content.push({
-        type: "text",
-        text: text.slice(lastIndex, m.index),
-      });
-    }
-
-    // マッチしたコンテンツ
-    if (m.type === "wikiLink") {
-      content.push({
-        type: "text",
-        text: `[[${m.text}]]`,
-        marks: [
-          {
-            type: "wikiLink",
-            attrs: {
-              title: m.text,
-              exists: false, // 新規生成なので存在しないとする
-            },
-          },
-        ],
-      });
-    } else if (m.type === "link") {
-      // 外部リンク
-      content.push({
-        type: "text",
-        text: m.text,
-        marks: [
-          {
-            type: "link",
-            attrs: {
-              href: m.url,
-              target: "_blank",
-              rel: "noopener noreferrer",
-            },
-          },
-        ],
-      });
-    } else if (m.type === "bold") {
-      content.push({
-        type: "text",
-        text: m.text,
-        marks: [{ type: "bold" }],
-      });
-    } else if (m.type === "italic") {
-      content.push({
-        type: "text",
-        text: m.text,
-        marks: [{ type: "italic" }],
-      });
-    }
-
-    lastIndex = m.index + m.length;
-  }
-
-  // 残りのテキスト
-  if (lastIndex < text.length) {
-    content.push({
-      type: "text",
-      text: text.slice(lastIndex),
-    });
-  }
-
-  // コンテンツがない場合は空のテキストを返す
-  if (content.length === 0 && text.length > 0) {
-    content.push({
-      type: "text",
-      text: text,
-    });
-  }
-
-  return content;
-}
+export { convertMarkdownToTiptapContent } from "./markdownToTiptap";
