+++ src/pages/NotePageView.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect } from "react";
+import React, { useCallback, useEffect, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { ArrowLeft } from "lucide-react";
 import Header from "@/components/layout/Header";
@@ -10,6 +10,8 @@ import { useAuth } from "@/hooks/useAuth";
 import { useCollaboration } from "@/hooks/useCollaboration";
 import { ContentWithAIChat } from "@/components/ai-chat/ContentWithAIChat";
 import { useAIChatContext } from "@/contexts/AIChatContext";
+import type { UseCollaborationReturn } from "@/lib/collaboration/types";
+import type { Page } from "@/types/page";
 
 function canEditPage(
   access: { canEdit?: boolean; canView?: boolean } | undefined,
@@ -19,6 +21,58 @@ function canEditPage(
   if (!access?.canView) return false;
   if (access.canEdit) return true;
   return Boolean(userId && page?.ownerUserId && page.ownerUserId === userId);
+}
+
+/** key={page.id} でページ切替時にリセット。editorContent の初期値を page.content から取得。 */
+function NotePageEditorEditable({
+  page,
+  collaboration,
+  isCollaborationEnabled,
+}: {
+  page: Page;
+  collaboration: UseCollaborationReturn;
+  isCollaborationEnabled: boolean;
+}) {
+  const [editorContent, setEditorContent] = useState(page.content ?? "");
+  const { setPageContext, contentAppendHandlerRef } = useAIChatContext();
+
+  useEffect(() => {
+    setPageContext({
+      type: "editor",
+      pageId: page.id,
+      pageTitle: page.title,
+      pageContent: editorContent.slice(0, 3000),
+      pageFullContent: editorContent,
+    });
+    return () => setPageContext(null);
+  }, [page.id, page.title, editorContent, setPageContext]);
+
+  useEffect(() => {
+    contentAppendHandlerRef.current = setEditorContent;
+    return () => {
+      contentAppendHandlerRef.current = null;
+    };
+  }, [contentAppendHandlerRef]);
+
+  return (
+    <ContentWithAIChat>
+      <PageEditorContent
+        content={editorContent}
+        title={page.title}
+        sourceUrl={page.sourceUrl}
+        currentPageId={page.id}
+        pageId={page.id}
+        isNewPage={false}
+        isWikiGenerating={false}
+        isReadOnly={false}
+        showLinkedPages={false}
+        showToolbar
+        onContentChange={setEditorContent}
+        onContentError={() => undefined}
+        collaboration={isCollaborationEnabled ? collaboration : undefined}
+      />
+    </ContentWithAIChat>
+  );
 }
 
 const NotePageView: React.FC = () => {
@@ -56,21 +110,6 @@ const NotePageView: React.FC = () => {
     enabled: isCollaborationEnabled,
     mode: "collaborative",
   });
-
-  // canEdit時のみAIチャットにページコンテキストを設定
-  const { setPageContext } = useAIChatContext();
-  useEffect(() => {
-    if (canEdit && page) {
-      setPageContext({
-        type: "editor",
-        pageId: page.id,
-        pageTitle: page.title,
-        pageContent: page.content?.slice(0, 3000) ?? "",
-        pageFullContent: page.content ?? "",
-      });
-    }
-    return () => setPageContext(null);
-  }, [canEdit, page, setPageContext]);
 
   const isLoading = isNoteLoading || isPageLoading;
   const isNotFound = !note || !access?.canView || !page;
@@ -114,26 +153,15 @@ const NotePageView: React.FC = () => {
       </div>
 
       {canEdit ? (
-        <ContentWithAIChat>
-          <PageEditorContent
-            content={page.content}
-            title={page.title}
-            sourceUrl={page.sourceUrl}
-            currentPageId={page.id}
-            pageId={page.id}
-            isNewPage={false}
-            isWikiGenerating={false}
-            isReadOnly={!canEdit}
-            showLinkedPages={false}
-            showToolbar={canEdit}
-            onContentChange={() => undefined}
-            onContentError={() => undefined}
-            collaboration={isCollaborationEnabled ? collaboration : undefined}
-          />
-        </ContentWithAIChat>
+        <NotePageEditorEditable
+          key={page.id}
+          page={page}
+          collaboration={collaboration}
+          isCollaborationEnabled={isCollaborationEnabled}
+        />
       ) : (
         <PageEditorContent
-          content={page.content}
+          content={page?.content ?? ""}
           title={page.title}
           sourceUrl={page.sourceUrl}
           currentPageId={page.id}
